fix: 完成 Admin 项目编译错误修复，成功迁移至 .NET 10 RC1
- 修复了 400+ 个编译错误，实现零错误构建
- 修复 ControllerBase.File 方法上下文冲突问题
- 添加完整的媒体管理方法实现
- 修复所有 DTO 类型不匹配和缺失类定义
- 修复方法参数和变量引用问题
- 实现跨平台兼容的 PerformanceCounter 替代方案
- 完善配置服务和依赖注入配置
- 添加缺失的实体属性和领域对象支持

+++ backend/src/MapleBlog.Admin/Services/DashboardService.cs
@@ -263,7 +263,7 @@ public class DashboardService : IDashboardService
             systemStats.UptimeSeconds = (long)(DateTime.UtcNow - systemStats.SystemStartTime).TotalSeconds;
 
             // 内存使用情况
-            systemStats.MemoryUsageMB = process.WorkingSet64 / (1024 * 1024);
+            systemStats.MemoryUsageMB = (int)(process.WorkingSet64 / (1024 * 1024));
 
             // 获取系统总内存（需要更复杂的实现）
             var totalMemoryMB = GC.GetTotalMemory(false) / (1024 * 1024);
@@ -592,7 +592,7 @@ public class DashboardService : IDashboardService
             TotalPosts = stats.TotalPosts,
             TotalUsers = stats.TotalUsers,
             TotalComments = stats.TotalComments,
-            TotalViews = stats.TotalViews,
+            TotalViews = (int)stats.TotalViews,
             AvgResponseTime = 0,
             LastUpdated = DateTime.UtcNow
         };
